OMS-N: Make asset manager cookies opaque [3/14]
Make asset manager cookies true first-class opaque handles.

When adding resources to an asset manager a handle (cookie) is returned
to the caller. The value of this handle used to have a one-to-one
correlation with an underlying array. This implementation detail had
leaked outside the asset manager boundaries and callers were relying on
it as an undocumented invariant, which prevented any modifications of
the asset manager except for the addition of more resources.

Change how cookies are treated by callers in order to unlock the
possibility to modify resources within an asset manager.

Known caveat: this commit will introduce some recurrent overhead. It
would probably make more sense to cache the retrieved information as
part of makeStringBlocks.

Bug: 31052947

AOSP Change-Id: I87b342a385dedeaa8b4a31cfe207cd33e31c7630

Change-Id: If7c4ad89d36dae932a6222f0ff5fc957af670d0b
Signed-off-by: KS-TheWiz <ksthewiz@gmail.com>