Allow unlocked USB data access (1/2)
by @KreAch3R

This bypasses the marshmallow "regression" of blocking and resetting USB data access
after disconnecting. It also allows saving the default configuration and it survives a reboot.
Originally, this "feature" was added by Google for security reasons
so it should be advised to use the below setting under caution.

Signed-off-by: KS-TheWiz <ksthewiz@gmail.com>