OMS-N: Officially announce the rebase completion to Nougat [14/14]
This commit cleans up the possible issues with the current iteration
of the Overlay Manager Service and it's free-compatibility with the
Android Open Source Project based ROMs.

This commit cleans up the commit to support downloaded overlays, but
adapted to Android Nougat's new code.

NOTE: enforceCodePolicy() has been added in Nougat to enforce checks on
hasCode manifest metadata tags to ensure that APKs must declare that it
wants an APK to be installed with no specified code. However this is
rather harmful to us since our overlay APKs install as non-code based
packages, thus tripping up the enforcement, ultimately requiring to be
disabled.

Change-Id: I077a11c155b41c1463d2cabdf7dcecd63d038620
Signed-off-by: KS-TheWiz <ksthewiz@gmail.com>