Fix issue of QS button color
The issue is that QS button show ON-color in OFF state, after
adjusting location or volume settings is re-allowed by device policy.

When QSTile is disabledByPolicy, Colorfilter is applied to Icon's
ImageView. Applying ColorFilter to ImageView changes the alpha of
Drawable with the value ImageView's alpha(mAlpha). That is,
the Button drawable could lose its intrinsic alpha after applying
ColorFilter to ImageView. So after ColorFilter is cleared, OFF state
Icon drawable's alpha unexpectedly remains 255. As a result,
QS-button shows ON-color in OFF-state.

The idea of fixing this issue is that always set ImageView's alpha
with the intrinsic alpha of current QS Drawable. By doing this,
Applying ColorFilter cannot changes the intrinsic alpha of Drawable.

Bug:29588731
Change-Id: I8371725b6f49ae9753cb8c3cce79dd668f6d1c7d
Signed-off-by: KS-TheWiz <ksthewiz@gmail.com>